backend/requirements.txt
# Flask application requirements for ResNet50V2 model
# Requires Python 3.9+ for optimal performance and security

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
numpy==1.24.3
tensorflow==2.13.0
requests==2.31.0
oauthlib==3.2.2
scipy==1.11.4
h5py==3.10.0
pymongo==4.5.0
flask-pymongo==2.3.0
dnspython==2.4.2
google-generativeai==0.3.2

# Google Cloud Vision API (fallback) with compatible protobuf
google-cloud-vision==3.4.4
protobuf==4.25.1

# Firebase Admin SDK
firebase-admin==6.0.1
